move addIncludePath from trait to class
