+++ libraries/src/MVC/Model/LeagcyModelLoaderTrait.php
@@ -23,56 +23,6 @@ use Joomla\CMS\Table\Table;
  */
 trait LeagcyModelLoaderTrait
 {
-	/**
-	 * Add a directory where \JModelLegacy should search for models. You may
-	 * either pass a string or an array of directories.
-	 *
-	 * @param   mixed   $path    A path or array[sting] of paths to search.
-	 * @param   string  $prefix  A prefix for models.
-	 *
-	 * @return  array  An array with directory elements. If prefix is equal to '', all directories are returned.
-	 *
-	 * @since       3.0
-	 * @deprecated  5.0 See getInstance
-	 */
-	public static function addIncludePath($path = '', $prefix = '')
-	{
-		static $paths;
-
-		if (!isset($paths))
-		{
-			$paths = array();
-		}
-
-		if (!isset($paths[$prefix]))
-		{
-			$paths[$prefix] = array();
-		}
-
-		if (!isset($paths['']))
-		{
-			$paths[''] = array();
-		}
-
-		if (!empty($path))
-		{
-			foreach ((array) $path as $includePath)
-			{
-				if (!\in_array($includePath, $paths[$prefix]))
-				{
-					array_unshift($paths[$prefix], Path::clean($includePath));
-				}
-
-				if (!\in_array($includePath, $paths['']))
-				{
-					array_unshift($paths[''], Path::clean($includePath));
-				}
-			}
-		}
-
-		return $paths[$prefix];
-	}
-
 
 	/**
 	 * Create the filename for a resource
